[ Round robin queue ordering ]
Each worker keeps its own ordered list of queues.  When it encounters a job on that queue, it rotates the queue so that the next item on the list is the first processed on the next reserve.  If no jobs are found, the list is untouched.

This solves a problem with the form of alpha-ordering or initializer-provided ordering that was previously implemented. If an early queue suddenly had a whole lot of work dumped on it, all other queues that a particular worker was watching would completely stall until the work was burned off.

With the round-robin queue strategy in this commit, the best way to prioritize a particular task would be to put it in a short/rarely-used/new queue. The attention that a particular queue receives is completely decoupled from how much work is on other queues.